Wayland refactor: Put handlers for windows in windows.rs, and separate WaylandState
